requirements.txt
Pillow~=10.2.0